REFACTOR: 커스텀 훅 분리 및 App.tsx 코드 정리
